[feature] repository dan handler user

+++ internal/domain/wallet.go
@@ -3,5 +3,5 @@ package domain
 type Wallet struct {
 	ID      int   `json:"id"`
 	UserId  User  `json:"user_id"`
-	Balance int64 `json:"balance"`
+	Balance float64 `json:"balance"`
 }
